+++ tests/brc-20/brc20.test.ts
@@ -1,77 +1,18 @@
 import { runMigrations } from '@hirosystems/api-toolkit';
 import { buildApiServer } from '../../src/api/init';
-import { Brc20ActivityResponse, Brc20TokenResponse } from '../../src/api/schemas';
-import { BRC20_SELF_MINT_ACTIVATION_BLOCK, brc20FromInscription } from '../../src/pg/brc20/helpers';
 import { MIGRATIONS_DIR, PgStore } from '../../src/pg/pg-store';
-import { DbLocationTransferType, InscriptionRevealData } from '../../src/pg/types';
 import {
+  BRC20_GENESIS_BLOCK,
+  BRC20_SELF_MINT_ACTIVATION_BLOCK,
   TestChainhookPayloadBuilder,
   TestFastifyServer,
-  brc20Reveal,
-  incrementing,
-  randomHash,
+  deployAndMintPEPE,
   rollBack,
 } from '../helpers';
-import { BRC20_GENESIS_BLOCK } from '../../src/pg/brc20/brc20-pg-store';
 
 describe('BRC-20', () => {
   let db: PgStore;
   let fastify: TestFastifyServer;
-
-  const deployAndMintPEPE = async (address: string) => {
-    await db.updateInscriptions(
-      new TestChainhookPayloadBuilder()
-        .apply()
-        .block({
-          height: BRC20_GENESIS_BLOCK,
-          hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-        })
-        .transaction({
-          hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-        })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'deploy',
-              tick: 'PEPE',
-              max: '250000',
-            },
-            number: 0,
-            ordinal_number: 0,
-            tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-            address: address,
-          })
-        )
-        .build()
-    );
-    await db.updateInscriptions(
-      new TestChainhookPayloadBuilder()
-        .apply()
-        .block({
-          height: BRC20_GENESIS_BLOCK + 1,
-          hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-        })
-        .transaction({
-          hash: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-        })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'mint',
-              tick: 'PEPE',
-              amt: '10000',
-            },
-            number: 1,
-            ordinal_number: 1,
-            tx_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-            address: address,
-          })
-        )
-        .build()
-    );
-  };
 
   beforeEach(async () => {
     await runMigrations(MIGRATIONS_DIR, 'up');
@@ -83,600 +24,6 @@ describe('BRC-20', () => {
     await fastify.close();
     await db.close();
     await runMigrations(MIGRATIONS_DIR, 'down');
-  });
-
-  describe('token standard validation', () => {
-    const testInsert = (json: any, block_height: number = 830000): InscriptionRevealData => {
-      const content = Buffer.from(JSON.stringify(json), 'utf-8');
-      return {
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          number: 0,
-          classic_number: 0,
-          mime_type: 'application/json',
-          content_type: 'application/json',
-          content_length: content.length,
-          content: `0x${content.toString('hex')}`,
-          fee: '200',
-          curse_type: null,
-          sat_ordinal: '2000000',
-          sat_rarity: 'common',
-          sat_coinbase_height: 110,
-          recursive: false,
-          metadata: null,
-          parent: null,
-        },
-        recursive_refs: [],
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height,
-          block_hash: '00000000000000000002c5c0aba96f981642a6dca109e6b3564925c21a98aa3e',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          tx_index: 0,
-          address: 'bc1pdjd6q33l0ca9nuudu2hr5qrs9u5dt6nl0z7fvu8kv4y8w4fzdpysc80028',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          prev_output: null,
-          prev_offset: null,
-          value: '9999',
-          transfer_type: DbLocationTransferType.transferred,
-          block_transfer_index: null,
-          timestamp: 1091091019,
-        },
-      };
-    };
-
-    test('ignores incorrect MIME type', () => {
-      const content = Buffer.from(
-        JSON.stringify({
-          p: 'brc-20',
-          op: 'deploy',
-          tick: 'PEPE',
-          max: '21000000',
-        }),
-        'utf-8'
-      );
-      const insert: InscriptionRevealData = {
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          number: 0,
-          classic_number: 0,
-          mime_type: 'foo/bar',
-          content_type: 'foo/bar;x=1',
-          content_length: content.length,
-          content: `0x${content.toString('hex')}`,
-          fee: '200',
-          curse_type: null,
-          sat_ordinal: '2000000',
-          sat_rarity: 'common',
-          sat_coinbase_height: 110,
-          recursive: false,
-          metadata: null,
-          parent: null,
-        },
-        recursive_refs: [],
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height: 830000,
-          block_hash: '00000000000000000002c5c0aba96f981642a6dca109e6b3564925c21a98aa3e',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          tx_index: 0,
-          address: 'bc1pdjd6q33l0ca9nuudu2hr5qrs9u5dt6nl0z7fvu8kv4y8w4fzdpysc80028',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          prev_output: null,
-          prev_offset: null,
-          value: '9999',
-          transfer_type: DbLocationTransferType.transferred,
-          block_transfer_index: null,
-          timestamp: 1091091019,
-        },
-      };
-      expect(brc20FromInscription(insert)).toBeUndefined();
-      insert.inscription.content_type = 'application/json';
-      insert.inscription.mime_type = 'application/json';
-      expect(brc20FromInscription(insert)).not.toBeUndefined();
-      insert.inscription.content_type = 'text/plain;charset=utf-8';
-      insert.inscription.mime_type = 'text/plain';
-      expect(brc20FromInscription(insert)).not.toBeUndefined();
-    });
-
-    test('ignores invalid JSON', () => {
-      const content = Buffer.from(
-        '{"p": "brc-20", "op": "deploy", "tick": "PEPE", "max": "21000000"',
-        'utf-8'
-      );
-      const insert: InscriptionRevealData = {
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          number: 0,
-          classic_number: 0,
-          mime_type: 'application/json',
-          content_type: 'application/json',
-          content_length: content.length,
-          content: `0x${content.toString('hex')}`,
-          fee: '200',
-          curse_type: null,
-          sat_ordinal: '2000000',
-          sat_rarity: 'common',
-          sat_coinbase_height: 110,
-          recursive: false,
-          metadata: null,
-          parent: null,
-        },
-        recursive_refs: [],
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height: 830000,
-          block_hash: '00000000000000000002c5c0aba96f981642a6dca109e6b3564925c21a98aa3e',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          tx_index: 0,
-          address: 'bc1pdjd6q33l0ca9nuudu2hr5qrs9u5dt6nl0z7fvu8kv4y8w4fzdpysc80028',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          prev_output: null,
-          prev_offset: null,
-          value: '9999',
-          transfer_type: DbLocationTransferType.transferred,
-          block_transfer_index: null,
-          timestamp: 1091091019,
-        },
-      };
-      expect(brc20FromInscription(insert)).toBeUndefined();
-    });
-
-    test('ignores inscriptions spent as fees', () => {
-      const content = Buffer.from(
-        JSON.stringify({
-          p: 'brc-20',
-          op: 'deploy',
-          tick: 'PEPE',
-          max: '21000000',
-        }),
-        'utf-8'
-      );
-      const insert: InscriptionRevealData = {
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          number: 0,
-          classic_number: 0,
-          mime_type: 'application/json',
-          content_type: 'application/json',
-          content_length: content.length,
-          content: `0x${content.toString('hex')}`,
-          fee: '200',
-          curse_type: null,
-          sat_ordinal: '2000000',
-          sat_rarity: 'common',
-          sat_coinbase_height: 110,
-          recursive: false,
-          metadata: null,
-          parent: null,
-        },
-        recursive_refs: [],
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height: 830000,
-          block_hash: '00000000000000000002c5c0aba96f981642a6dca109e6b3564925c21a98aa3e',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          tx_index: 0,
-          address: '',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          prev_output: null,
-          prev_offset: null,
-          value: '0',
-          transfer_type: DbLocationTransferType.spentInFees,
-          block_transfer_index: null,
-          timestamp: 1091091019,
-        },
-      };
-      expect(brc20FromInscription(insert)).toBeUndefined();
-    });
-
-    test('ignores burnt inscriptions', () => {
-      const content = Buffer.from(
-        JSON.stringify({
-          p: 'brc-20',
-          op: 'deploy',
-          tick: 'PEPE',
-          max: '21000000',
-        }),
-        'utf-8'
-      );
-      const insert: InscriptionRevealData = {
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          number: 0,
-          classic_number: 0,
-          mime_type: 'application/json',
-          content_type: 'application/json',
-          content_length: content.length,
-          content: `0x${content.toString('hex')}`,
-          fee: '200',
-          curse_type: null,
-          sat_ordinal: '2000000',
-          sat_rarity: 'common',
-          sat_coinbase_height: 110,
-          recursive: false,
-          metadata: null,
-          parent: null,
-        },
-        recursive_refs: [],
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height: 830000,
-          block_hash: '00000000000000000002c5c0aba96f981642a6dca109e6b3564925c21a98aa3e',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          tx_index: 0,
-          address: '',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          prev_output: null,
-          prev_offset: null,
-          value: '1000',
-          transfer_type: DbLocationTransferType.burnt,
-          block_transfer_index: null,
-          timestamp: 1091091019,
-        },
-      };
-      expect(brc20FromInscription(insert)).toBeUndefined();
-    });
-
-    test('ignores incorrect p field', () => {
-      const insert = testInsert({
-        p: 'brc20', // incorrect
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert)).toBeUndefined();
-    });
-
-    test('ignores incorrect op field', () => {
-      const insert = testInsert({
-        p: 'brc-20',
-        op: 'deploi', // incorrect
-        tick: 'PEPE',
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert)).toBeUndefined();
-    });
-
-    test('tick must be 4 or 5 bytes wide', () => {
-      const insert = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPETESTER', // more than 4 bytes
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert)).toBeUndefined();
-      const insert2 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'Pe P', // valid
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert2)).not.toBeUndefined();
-      const insert3 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: '🤬😉', // more than 4 bytes
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert3)).toBeUndefined();
-      const insert4 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'X', // less than 4 bytes
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert4)).toBeUndefined();
-    });
-
-    test('deploy self_mint tick must be 5 bytes wide', () => {
-      const insert = testInsert(
-        {
-          p: 'brc-20',
-          op: 'deploy',
-          tick: '$PEPE', // 5 bytes
-          max: '21000000',
-          self_mint: 'true',
-        },
-        840000
-      );
-      expect(brc20FromInscription(insert)).not.toBeUndefined();
-      const insert2 = testInsert(
-        {
-          p: 'brc-20',
-          op: 'deploy',
-          tick: '$PEPE', // 5 bytes but no self_mint
-          max: '21000000',
-        },
-        840000
-      );
-      expect(brc20FromInscription(insert2)).toBeUndefined();
-      const insert4 = testInsert(
-        {
-          p: 'brc-20',
-          op: 'deploy',
-          tick: '$PEPE', // Correct but earlier than activation
-          max: '21000000',
-          self_mint: 'true',
-        },
-        820000
-      );
-      expect(brc20FromInscription(insert4)).toBeUndefined();
-    });
-
-    test('all fields must be strings', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: 21000000,
-      });
-      expect(brc20FromInscription(insert1)).toBeUndefined();
-      const insert1a = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: 300,
-      });
-      expect(brc20FromInscription(insert1a)).toBeUndefined();
-      const insert1b = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '300',
-        dec: 2,
-      });
-      expect(brc20FromInscription(insert1b)).toBeUndefined();
-      const insert2 = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: 2,
-      });
-      expect(brc20FromInscription(insert2)).toBeUndefined();
-      const insert3 = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: 2,
-      });
-      expect(brc20FromInscription(insert3)).toBeUndefined();
-    });
-
-    test('ignores empty strings', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: '',
-        max: '21000000',
-      });
-      expect(brc20FromInscription(insert1)).toBeUndefined();
-      const insert1a = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '',
-      });
-      expect(brc20FromInscription(insert1a)).toBeUndefined();
-      const insert1b = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '',
-      });
-      expect(brc20FromInscription(insert1b)).toBeUndefined();
-      const insert1c = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '200',
-        dec: '',
-      });
-      expect(brc20FromInscription(insert1c)).toBeUndefined();
-      const insert2 = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: '',
-      });
-      expect(brc20FromInscription(insert2)).toBeUndefined();
-      const insert2a = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: '',
-      });
-      expect(brc20FromInscription(insert2a)).toBeUndefined();
-      const insert3 = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: '',
-      });
-      expect(brc20FromInscription(insert3)).toBeUndefined();
-      const insert3a = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: '',
-      });
-      expect(brc20FromInscription(insert3a)).toBeUndefined();
-    });
-
-    test('numeric strings must not be zero', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '0',
-      });
-      expect(brc20FromInscription(insert1)).toBeUndefined();
-      const insert1b = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '0.0',
-      });
-      expect(brc20FromInscription(insert1b)).toBeUndefined();
-      const insert1c = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '200',
-        dec: '0',
-      });
-      // `dec` can have a value of 0
-      expect(brc20FromInscription(insert1c)).not.toBeUndefined();
-      const insert1d = testInsert(
-        {
-          p: 'brc-20',
-          op: 'deploy',
-          tick: '$PEPE',
-          max: '0', // self mints can be max 0
-          self_mint: 'true',
-        },
-        840000
-      );
-      expect(brc20FromInscription(insert1d)).not.toBeUndefined();
-      const insert2a = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: '0',
-      });
-      expect(brc20FromInscription(insert2a)).toBeUndefined();
-      const insert3a = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: '.0000',
-      });
-      expect(brc20FromInscription(insert3a)).toBeUndefined();
-    });
-
-    test('numeric fields are not stripped/trimmed', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: ' 200  ',
-      });
-      expect(brc20FromInscription(insert1)).toBeUndefined();
-      const insert1b = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '+10000',
-      });
-      expect(brc20FromInscription(insert1b)).toBeUndefined();
-      const insert1c = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '200',
-        dec: '   0 ',
-      });
-      expect(brc20FromInscription(insert1c)).toBeUndefined();
-      const insert2a = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: '.05 ',
-      });
-      expect(brc20FromInscription(insert2a)).toBeUndefined();
-      const insert3a = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: '-25.00',
-      });
-      expect(brc20FromInscription(insert3a)).toBeUndefined();
-    });
-
-    test('max value of dec is 18', () => {
-      const insert1c = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '200',
-        dec: '20',
-      });
-      expect(brc20FromInscription(insert1c)).toBeUndefined();
-    });
-
-    test('max value of any numeric field is uint64_max', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '18446744073709551999',
-      });
-      expect(brc20FromInscription(insert1)).toBeUndefined();
-      const insert1b = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '21000000',
-        lim: '18446744073709551999',
-      });
-      expect(brc20FromInscription(insert1b)).toBeUndefined();
-      const insert2a = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: '18446744073709551999',
-      });
-      expect(brc20FromInscription(insert2a)).toBeUndefined();
-      const insert3a = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: '18446744073709551999',
-      });
-      expect(brc20FromInscription(insert3a)).toBeUndefined();
-    });
-
-    test('valid JSONs can have additional properties', () => {
-      const insert1 = testInsert({
-        p: 'brc-20',
-        op: 'deploy',
-        tick: 'PEPE',
-        max: '200',
-        foo: 'bar',
-        test: 1,
-      });
-      expect(brc20FromInscription(insert1)).not.toBeUndefined();
-      const insert2a = testInsert({
-        p: 'brc-20',
-        op: 'mint',
-        tick: 'PEPE',
-        amt: '5',
-        foo: 'bar',
-        test: 1,
-      });
-      expect(brc20FromInscription(insert2a)).not.toBeUndefined();
-      const insert3a = testInsert({
-        p: 'brc-20',
-        op: 'transfer',
-        tick: 'PEPE',
-        amt: '25',
-        foo: 'bar',
-        test: 1,
-      });
-      expect(brc20FromInscription(insert3a)).not.toBeUndefined();
-    });
   });
 
   describe('deploy', () => {
@@ -692,20 +39,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: 'PEPE',
+              max: '21000000',
+              lim: '1000',
+              dec: '18',
               address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
+              self_mint: false,
+            },
+          })
           .build()
       );
       const response1 = await fastify.inject({
@@ -734,44 +78,6 @@ describe('BRC-20', () => {
       ]);
     });
 
-    test('deploy with self_mint is ignored before activation height', async () => {
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-            timestamp: 1677811111,
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '21000000',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=$PEPE`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const responseJson1 = response1.json();
-      expect(responseJson1.total).toBe(0);
-    });
-
     test('deploy with self_mint is saved', async () => {
       await db.updateInscriptions(
         new TestChainhookPayloadBuilder()
@@ -784,21 +90,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '21000000',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: '$PEPE',
+              max: '21000000',
+              lim: '1000',
+              dec: '18',
               address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
+              self_mint: true,
+            },
+          })
           .build()
       );
       const response1 = await fastify.inject({
@@ -824,226 +126,6 @@ describe('BRC-20', () => {
         tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
       });
     });
-
-    test('ignores deploys for existing token', async () => {
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '000000000000000000021a0207fa97024506baaa74396822fb0a07ac20e70148',
-          })
-          .transaction({
-            hash: '3f8067a6e9b45308b5a090c2987feeb2d08cbaf814ef2ffabad7c381b62f5f7e',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '19000000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '3f8067a6e9b45308b5a090c2987feeb2d08cbaf814ef2ffabad7c381b62f5f7e',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=PEPE`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const responseJson1 = response1.json();
-      expect(responseJson1.total).toBe(1);
-      expect(responseJson1.results).toStrictEqual([
-        {
-          address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-          block_height: BRC20_GENESIS_BLOCK,
-          decimals: 18,
-          id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          max_supply: '21000000.000000000000000000',
-          mint_limit: null,
-          number: 0,
-          ticker: 'PEPE',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          deploy_timestamp: 1677803510000,
-          minted_supply: '0.000000000000000000',
-          tx_count: 1,
-          self_mint: false,
-        },
-      ]);
-    });
-
-    test('ignores case insensitive deploy for existing token', async () => {
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '000000000000000000021a0207fa97024506baaa74396822fb0a07ac20e70148',
-          })
-          .transaction({
-            hash: '3f8067a6e9b45308b5a090c2987feeb2d08cbaf814ef2ffabad7c381b62f5f7e',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'pepe',
-                max: '19000000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '3f8067a6e9b45308b5a090c2987feeb2d08cbaf814ef2ffabad7c381b62f5f7e',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=PEPE`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const responseJson1 = response1.json();
-      expect(responseJson1.total).toBe(1);
-      expect(responseJson1.results).toStrictEqual([
-        {
-          address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-          block_height: BRC20_GENESIS_BLOCK,
-          decimals: 18,
-          id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          max_supply: '21000000.000000000000000000',
-          mint_limit: null,
-          number: 0,
-          ticker: 'PEPE',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          deploy_timestamp: 1677803510000,
-          minted_supply: '0.000000000000000000',
-          tx_count: 1,
-          self_mint: false,
-        },
-      ]);
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=pepe`, // Lowercase
-      });
-      expect(response2.statusCode).toBe(200);
-      const responseJson2 = response2.json();
-      expect(responseJson2.total).toBe(1);
-      expect(responseJson2.results).toStrictEqual([
-        {
-          address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-          block_height: BRC20_GENESIS_BLOCK,
-          decimals: 18,
-          id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          max_supply: '21000000.000000000000000000',
-          mint_limit: null,
-          number: 0,
-          ticker: 'PEPE',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          deploy_timestamp: 1677803510000,
-          minted_supply: '0.000000000000000000',
-          tx_count: 1,
-          self_mint: false,
-        },
-      ]);
-    });
-
-    test('ignores deploy from classic cursed inscription', async () => {
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              classic_number: -1,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
-          .build()
-      );
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=PEPE`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const responseJson1 = response1.json();
-      expect(responseJson1.total).toBe(0);
-      expect(responseJson1.results).toHaveLength(0);
-    });
   });
 
   describe('mint', () => {
@@ -1059,20 +141,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: 'PEPE',
+              max: '21000000',
+              lim: '250000',
+              dec: '18',
+              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
+              self_mint: false,
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -1085,20 +164,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-            })
-          )
+          .brc20({
+            mint: {
+              tick: 'PEPE',
+              amt: '250000',
+              inscription_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99fi0',
+              address,
+            },
+          })
           .build()
       );
 
@@ -1129,20 +202,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'pepe',
-                amt: '100000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            })
-          )
+          .brc20({
+            mint: {
+              tick: 'PEPE',
+              amt: '100000',
+              inscription_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8beci0',
+              address,
+            },
+          })
           .build()
       );
 
@@ -1188,21 +255,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '21000000',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: '$PEPE',
+              max: '21000000',
+              lim: '21000000',
+              dec: '18',
+              address,
+              self_mint: true,
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -1215,21 +278,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-              parent: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99fi0',
+              tick: '$PEPE',
+              address,
+              amt: '250000',
+            },
+          })
           .build()
       );
 
@@ -1260,21 +316,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$pepe',
-                amt: '100000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
+          .brc20({
+            mint: {
+              inscription_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8beci0',
+              tick: '$PEPE',
               address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              parent: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            })
-          )
+              amt: '100000',
+            },
+          })
           .build()
       );
 
@@ -1308,85 +357,6 @@ describe('BRC-20', () => {
       );
     });
 
-    test('self mints with invalid parent inscription are ignored', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_SELF_MINT_ACTIVATION_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '21000000',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_SELF_MINT_ACTIVATION_BLOCK + 1,
-            hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-          })
-          .transaction({
-            hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-              // no parent
-            })
-          )
-          .build()
-      );
-
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const responseJson1 = response1.json();
-      expect(responseJson1.total).toBe(0);
-
-      const response3 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/tokens?ticker=$PEPE`,
-      });
-      expect(response3.statusCode).toBe(200);
-      const responseJson3 = response3.json();
-      expect(responseJson3.total).toBe(1);
-      expect(responseJson3.results).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({ ticker: '$PEPE', minted_supply: '0.000000000000000000' }),
-        ])
-      );
-    });
-
     test('valid self mints for tokens with max 0 are saved', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
       await db.updateInscriptions(
@@ -1399,21 +369,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '0',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: '$PEPE',
+              max: '0',
+              lim: '250000',
+              dec: '18',
+              address,
+              self_mint: true,
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -1426,21 +392,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-              parent: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99fi0',
+              tick: '$PEPE',
+              address,
+              amt: '250000',
+            },
+          })
           .build()
       );
 
@@ -1471,21 +430,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$pepe',
-                amt: '100000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8bec',
-              address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              parent: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '7a1adbc3e93ddf8d7c4e0ba75aa11c98c431521dd850be8b955feedb716d8beci0',
+              tick: '$PEPE',
+              address,
+              amt: '100000',
+            },
+          })
           .build()
       );
 
@@ -1531,20 +483,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: 'PEPE',
+              max: '21000000',
+              lim: '21000000',
+              dec: '18',
+              address,
+              self_mint: false,
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -1557,20 +506,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99fi0',
+              tick: 'PEPE',
+              address,
+              amt: '250000',
+            },
+          })
           .build()
       );
       // Rollback
@@ -1584,20 +527,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '250000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99fi0',
+              tick: 'PEPE',
+              address,
+              amt: '250000',
+            },
+          })
           .build()
       );
 
@@ -1616,327 +553,12 @@ describe('BRC-20', () => {
       });
       expect(response3.json().token.minted_supply).toBe('0.000000000000000000');
     });
-
-    test('numbers should not have more decimal digits than "dec" of ticker', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '21000000',
-                dec: '1',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-          })
-          .transaction({
-            hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '250000.000', // Invalid decimal count
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response2.statusCode).toBe(200);
-      const responseJson2 = response2.json();
-      expect(responseJson2.total).toBe(0);
-      expect(responseJson2.results).toStrictEqual([]);
-    });
-
-    test('mint exceeds token supply', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '2500',
-                dec: '1',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-          })
-          .transaction({
-            hash: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '1000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-              address: address,
-            })
-          )
-          .transaction({
-            hash: '7e09bda2cba34bca648cca6d79a074940d39b6137150d3a3edcf80c0e01419a5',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '1000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: '7e09bda2cba34bca648cca6d79a074940d39b6137150d3a3edcf80c0e01419a5',
-              address: address,
-            })
-          )
-          .transaction({
-            hash: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '5000000000', // Exceeds supply
-              },
-              number: 3,
-              ordinal_number: 3,
-              tx_id: '8aec77f855549d98cb9fb5f35e02a03f9a2354fd05a5f89fc610b32c3b01f99f',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}?ticker=PEPE`,
-      });
-      expect(response2.statusCode).toBe(200);
-      const responseJson2 = response2.json();
-      expect(responseJson2.total).toBe(1);
-      expect(responseJson2.results).toStrictEqual([
-        {
-          available_balance: '2500.0', // Max capacity
-          overall_balance: '2500.0',
-          ticker: 'PEPE',
-          transferrable_balance: '0.0',
-        },
-      ]);
-
-      // No more mints allowed
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '000000000000000000001f14513d722146fddab04a1855665a5eca22df288c3c',
-          })
-          .transaction({
-            hash: 'bf7a3e1a0647ca88f6539119b2defaec302683704ea270b3302e709597643548',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '1000',
-              },
-              number: 4,
-              ordinal_number: 4,
-              tx_id: 'bf7a3e1a0647ca88f6539119b2defaec302683704ea270b3302e709597643548',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response3 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response3.statusCode).toBe(200);
-      const responseJson3 = response3.json();
-      expect(responseJson3).toStrictEqual(responseJson2);
-    });
-
-    test('ignores mint for non-existent token', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-          })
-          .transaction({
-            hash: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '1000',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response2.statusCode).toBe(200);
-      const responseJson2 = response2.json();
-      expect(responseJson2.total).toBe(0);
-      expect(responseJson2.results).toStrictEqual([]);
-    });
-
-    test('mint exceeds token mint limit', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK,
-            hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          })
-          .transaction({
-            hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: 'PEPE',
-                max: '2500',
-                dec: '1',
-                lim: '100',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 1,
-            hash: '0000000000000000000098d8f2663891d439f6bb7de230d4e9f6bcc2e85452bf',
-          })
-          .transaction({
-            hash: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: 'PEPE',
-                amt: '1000', // Greater than limit
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response2.statusCode).toBe(200);
-      const responseJson2 = response2.json();
-      expect(responseJson2.total).toBe(0);
-      expect(responseJson2.results).toStrictEqual([]);
-    });
   });
 
   describe('transfer', () => {
     test('available balance decreases on transfer inscription', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
+      await deployAndMintPEPE(db, address);
       await db.updateInscriptions(
         new TestChainhookPayloadBuilder()
           .apply()
@@ -1947,20 +569,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '2000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              tick: 'PEPE',
+              address,
+              amt: '2000',
+            },
+          })
           .build()
       );
 
@@ -1989,103 +605,9 @@ describe('BRC-20', () => {
       expect(json2.results[0].available_balance).toBe('10000.000000000000000000');
     });
 
-    test('transfer ignored if token not found', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '00000000000000000002b14f0c5dde0b2fc74d022e860696bd64f1f652756674',
-          })
-          .transaction({
-            hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'TEST', // Not found
-                amt: '2000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response.statusCode).toBe(200);
-      const json = response.json();
-      expect(json.total).toBe(1);
-      expect(json.results).toStrictEqual([
-        {
-          available_balance: '10000.000000000000000000',
-          overall_balance: '10000.000000000000000000',
-          ticker: 'PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-    });
-
-    test('cannot transfer more than available balance', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '00000000000000000002b14f0c5dde0b2fc74d022e860696bd64f1f652756674',
-          })
-          .transaction({
-            hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '5000000000', // More than was minted
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
-          .build()
-      );
-
-      const response = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response.statusCode).toBe(200);
-      const json = response.json();
-      expect(json.total).toBe(1);
-      expect(json.results).toStrictEqual([
-        {
-          available_balance: '10000.000000000000000000',
-          overall_balance: '10000.000000000000000000',
-          ticker: 'PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-    });
-
     test('multiple transfers in block', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
+      await deployAndMintPEPE(db, address);
       await db.updateInscriptions(
         new TestChainhookPayloadBuilder()
           .apply()
@@ -2096,37 +618,25 @@ describe('BRC-20', () => {
           .transaction({
             hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              tick: 'PEPE',
+              address,
+              amt: '9000',
+            },
+          })
           .transaction({
             hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '2000', // Will exceed available balance
-              },
-              number: 3,
-              ordinal_number: 3,
-              tx_id: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21baci0',
+              tick: 'PEPE',
+              address,
+              amt: '1000',
+            },
+          })
           .build()
       );
 
@@ -2150,7 +660,7 @@ describe('BRC-20', () => {
     test('send balance to address', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
       const address2 = '3QNjwPDRafjBm9XxJpshgk3ksMJh3TFxTU';
-      await deployAndMintPEPE(address);
+      await deployAndMintPEPE(db, address);
       await db.updateInscriptions(
         new TestChainhookPayloadBuilder()
           .apply()
@@ -2161,20 +671,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              tick: 'PEPE',
+              address,
+              amt: '9000',
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -2187,15 +691,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
           })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'transferred', value: address2 },
-            satpoint_pre_transfer:
-              'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-            satpoint_post_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
+          .brc20({
+            transfer_send: {
+              tick: 'PEPE',
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              amt: '9000',
+              sender_address: address,
+              receiver_address: address2,
+            },
           })
           .build()
       );
@@ -2260,21 +763,17 @@ describe('BRC-20', () => {
           .transaction({
             hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'deploy',
-                tick: '$PEPE',
-                max: '0',
-                self_mint: 'true',
-              },
-              number: 0,
-              ordinal_number: 0,
-              tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-              address: address,
-            })
-          )
+          .brc20({
+            deploy: {
+              inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+              tick: '$PEPE',
+              max: '0',
+              lim: '21000000',
+              dec: '18',
+              address,
+              self_mint: true,
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -2287,21 +786,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'mint',
-                tick: '$PEPE',
-                amt: '10000',
-              },
-              number: 1,
-              ordinal_number: 1,
-              tx_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0f',
-              address: address,
-              parent: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            })
-          )
+          .brc20({
+            mint: {
+              inscription_id: '3b55f624eaa4f8de6c42e0c490176b67123a83094384f658611faf7bfb85dd0fi0',
+              tick: '$PEPE',
+              address,
+              amt: '10000',
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -2314,20 +806,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: '$PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              tick: '$PEPE',
+              address,
+              amt: '9000',
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -2340,15 +826,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
           })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'transferred', value: address2 },
-            satpoint_pre_transfer:
-              'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-            satpoint_post_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
+          .brc20({
+            transfer_send: {
+              inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+              tick: '$PEPE',
+              amt: '9000',
+              sender_address: address,
+              receiver_address: address2,
+            },
           })
           .build()
       );
@@ -2381,255 +866,6 @@ describe('BRC-20', () => {
           available_balance: '9000.000000000000000000',
           overall_balance: '9000.000000000000000000',
           ticker: '$PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-    });
-
-    test('sending transfer as fee returns amount to sender', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '00000000000000000002b14f0c5dde0b2fc74d022e860696bd64f1f652756674',
-          })
-          .transaction({
-            hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 3,
-            hash: '00000000000000000003feae13d107f0f2c4fb4dd08fb2a8b1ab553512e77f03',
-          })
-          .transaction({
-            hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
-          })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'spent_in_fees', value: '' },
-            satpoint_pre_transfer:
-              'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-            satpoint_post_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
-          })
-          .build()
-      );
-
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const json1 = response1.json();
-      expect(json1.total).toBe(1);
-      expect(json1.results).toStrictEqual([
-        {
-          available_balance: '10000.000000000000000000',
-          overall_balance: '10000.000000000000000000',
-          ticker: 'PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-    });
-
-    test('sending transfer to unspendable output does not return to sender', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '00000000000000000002b14f0c5dde0b2fc74d022e860696bd64f1f652756674',
-          })
-          .transaction({
-            hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 3,
-            hash: '00000000000000000003feae13d107f0f2c4fb4dd08fb2a8b1ab553512e77f03',
-          })
-          .transaction({
-            hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
-          })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'burnt' },
-            satpoint_pre_transfer:
-              'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-            satpoint_post_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
-          })
-          .build()
-      );
-
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const json1 = response1.json();
-      expect(json1.total).toBe(1);
-      expect(json1.results).toStrictEqual([
-        {
-          available_balance: '1000.000000000000000000',
-          overall_balance: '1000.000000000000000000',
-          ticker: 'PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-    });
-
-    test('cannot spend valid transfer twice', async () => {
-      const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      const address2 = '3QNjwPDRafjBm9XxJpshgk3ksMJh3TFxTU';
-      await deployAndMintPEPE(address);
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 2,
-            hash: '00000000000000000002b14f0c5dde0b2fc74d022e860696bd64f1f652756674',
-          })
-          .transaction({
-            hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-          })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '9000',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-              address: address,
-            })
-          )
-          .build()
-      );
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 3,
-            hash: '000000000000000000016ddf56d0fe72476165acee9500d48d3e2aaf8412f489',
-          })
-          .transaction({
-            hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
-          })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'transferred', value: address2 },
-            satpoint_pre_transfer:
-              'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-            satpoint_post_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
-          })
-          .build()
-      );
-      // Attempt to transfer the same inscription back to the original address to change its
-      // balance.
-      await db.updateInscriptions(
-        new TestChainhookPayloadBuilder()
-          .apply()
-          .block({
-            height: BRC20_GENESIS_BLOCK + 4,
-            hash: '00000000000000000003feae13d107f0f2c4fb4dd08fb2a8b1ab553512e77f03',
-          })
-          .transaction({
-            hash: '55bec906eadc9f5c120cc39555ba46e85e562eacd6217e4dd0b8552783286d0e',
-          })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'transferred', value: address },
-            satpoint_pre_transfer:
-              '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            satpoint_post_transfer:
-              '55bec906eadc9f5c120cc39555ba46e85e562eacd6217e4dd0b8552783286d0e:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
-          })
-          .build()
-      );
-
-      // Balances only reflect the first transfer.
-      const response1 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address}`,
-      });
-      expect(response1.statusCode).toBe(200);
-      const json1 = response1.json();
-      expect(json1.total).toBe(1);
-      expect(json1.results).toStrictEqual([
-        {
-          available_balance: '1000.000000000000000000',
-          overall_balance: '1000.000000000000000000',
-          ticker: 'PEPE',
-          transferrable_balance: '0.000000000000000000',
-        },
-      ]);
-
-      const response2 = await fastify.inject({
-        method: 'GET',
-        url: `/ordinals/brc-20/balances/${address2}`,
-      });
-      expect(response2.statusCode).toBe(200);
-      const json2 = response2.json();
-      expect(json2.total).toBe(1);
-      expect(json2.results).toStrictEqual([
-        {
-          available_balance: '9000.000000000000000000',
-          overall_balance: '9000.000000000000000000',
-          ticker: 'PEPE',
           transferrable_balance: '0.000000000000000000',
         },
       ]);
@@ -2637,7 +873,7 @@ describe('BRC-20', () => {
 
     test('explicit transfer to self restores balance correctly', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-      await deployAndMintPEPE(address);
+      await deployAndMintPEPE(db, address);
       const address2 = 'bc1ph8dp3lqhzpjphqcc3ucgsm7k3w4d74uwfpv8sv893kn3kpkqrdxqqy3cv6';
       await db.updateInscriptions(
         new TestChainhookPayloadBuilder()
@@ -2649,20 +885,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '825a25b64b5d99ca30e04e53cc9a3020412e1054eb2a7523eb075ddd6d983205',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '20',
-              },
-              number: 2,
-              ordinal_number: 2,
-              tx_id: '825a25b64b5d99ca30e04e53cc9a3020412e1054eb2a7523eb075ddd6d983205',
-              address: address,
-            })
-          )
+          .brc20({
+            transfer: {
+              inscription_id: '825a25b64b5d99ca30e04e53cc9a3020412e1054eb2a7523eb075ddd6d983205i0',
+              tick: 'PEPE',
+              address,
+              amt: '20',
+            },
+          })
           .build()
       );
       await db.updateInscriptions(
@@ -2675,15 +905,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '486815e61723d03af344e1256d7e0c028a8e9e71eb38157f4bf069eb94292ee1',
           })
-          .inscriptionTransferred({
-            ordinal_number: 2,
-            destination: { type: 'transferred', value: address2 },
-            satpoint_pre_transfer:
-              '825a25b64b5d99ca30e04e53cc9a3020412e1054eb2a7523eb075ddd6d983205:0:0',
-            satpoint_post_transfer:
-              '486815e61723d03af344e1256d7e0c028a8e9e71eb38157f4bf069eb94292ee1:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
+          .brc20({
+            transfer_send: {
+              inscription_id: '825a25b64b5d99ca30e04e53cc9a3020412e1054eb2a7523eb075ddd6d983205i0',
+              tick: 'PEPE',
+              amt: '20',
+              sender_address: address,
+              receiver_address: address2,
+            },
           })
           .build()
       );
@@ -2709,20 +938,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '09a812f72275892b4858880cf3821004a6e8885817159b340639afe9952ac053',
           })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'PEPE',
-                amt: '20',
-              },
-              number: 3,
-              ordinal_number: 3,
-              tx_id: '09a812f72275892b4858880cf3821004a6e8885817159b340639afe9952ac053',
+          .brc20({
+            transfer: {
+              inscription_id: '09a812f72275892b4858880cf3821004a6e8885817159b340639afe9952ac053i0',
+              tick: 'PEPE',
               address: address2,
-            })
-          )
+              amt: '20',
+            },
+          })
           .build()
       );
       response = await fastify.inject({
@@ -2747,15 +970,14 @@ describe('BRC-20', () => {
           .transaction({
             hash: '26c0c3acbb1c87e682ade86220ba06e649d7599ecfc49a71495f1bdd04efbbb4',
           })
-          .inscriptionTransferred({
-            ordinal_number: 3,
-            destination: { type: 'transferred', value: address2 },
-            satpoint_pre_transfer:
-              '486815e61723d03af344e1256d7e0c028a8e9e71eb38157f4bf069eb94292ee1:0:0',
-            satpoint_post_transfer:
-              '26c0c3acbb1c87e682ade86220ba06e649d7599ecfc49a71495f1bdd04efbbb4:0:0',
-            post_transfer_output_value: null,
-            tx_index: 0,
+          .brc20({
+            transfer_send: {
+              inscription_id: '09a812f72275892b4858880cf3821004a6e8885817159b340639afe9952ac053i0',
+              tick: 'PEPE',
+              amt: '20',
+              sender_address: address2,
+              receiver_address: address2,
+            },
           })
           .build()
       );
@@ -2774,1393 +996,11 @@ describe('BRC-20', () => {
     });
   });
 
-  describe('routes', () => {
-    describe('/brc-20/tokens', () => {
-      test('tokens endpoint', async () => {
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: BRC20_GENESIS_BLOCK })
-            .transaction({
-              hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-            })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number: 0,
-                ordinal_number: 0,
-                tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-                address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              })
-            )
-            .build()
-        );
-        const response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens/PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        expect(response.json()).toStrictEqual({
-          token: {
-            id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-            number: 0,
-            block_height: BRC20_GENESIS_BLOCK,
-            tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-            address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            ticker: 'PEPE',
-            max_supply: '21000000.000000000000000000',
-            mint_limit: null,
-            decimals: 18,
-            deploy_timestamp: 1677803510000,
-            minted_supply: '0.000000000000000000',
-            tx_count: 1,
-            self_mint: false,
-          },
-          supply: {
-            max_supply: '21000000.000000000000000000',
-            minted_supply: '0.000000000000000000',
-            holders: 0,
-          },
-        });
-      });
-
-      test('tokens filter by ticker prefix', async () => {
-        const inscriptionNumbers = incrementing(0);
-        const blockHeights = incrementing(BRC20_GENESIS_BLOCK);
-
-        let transferHash = randomHash();
-        let number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: transferHash,
-                address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              })
-            )
-            .build()
-        );
-
-        transferHash = randomHash();
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEER',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: transferHash,
-                address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              })
-            )
-            .build()
-        );
-
-        transferHash = randomHash();
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'ABCD',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: transferHash,
-                address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              })
-            )
-            .build()
-        );
-
-        transferHash = randomHash();
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'DCBA',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: transferHash,
-                address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-              })
-            )
-            .build()
-        );
-        const response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens?ticker=PE&ticker=AB`,
-        });
-        expect(response.statusCode).toBe(200);
-        const responseJson = response.json();
-        expect(responseJson.total).toBe(3);
-        expect(responseJson.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({ ticker: 'PEPE' }),
-            expect.objectContaining({ ticker: 'PEER' }),
-            expect.objectContaining({ ticker: 'ABCD' }),
-          ])
-        );
-      });
-
-      test('tokens using order_by tx_count', async () => {
-        // Setup
-        const inscriptionNumbers = incrementing(0);
-        const blockHeights = incrementing(BRC20_GENESIS_BLOCK);
-        const addressA = 'bc1q6uwuet65rm6xvlz7ztw2gvdmmay5uaycu03mqz';
-        const addressB = 'bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4';
-
-        // A deploys PEPE
-        let number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // A mints 10000 PEPE 10 times (will later be rolled back)
-        const pepeMints = [];
-        for (let i = 0; i < 10; i++) {
-          const txHash = randomHash();
-          number = inscriptionNumbers.next().value;
-          const payload = new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: txHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '10000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: txHash,
-                address: addressA,
-              })
-            )
-            .build();
-          pepeMints.push(payload);
-          await db.updateInscriptions(payload);
-        }
-
-        // B deploys ABCD
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'ABCD',
-                  max: '21000000',
-                },
-                number: number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressB,
-              })
-            )
-            .build()
-        );
-
-        // B mints 10000 ABCD
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'ABCD',
-                  amt: '10000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressB,
-              })
-            )
-            .build()
-        );
-
-        // B send 1000 ABCD to A
-        // (create inscription, transfer)
-        const txHashTransfer = randomHash();
-        number = inscriptionNumbers.next().value;
-        const payloadTransfer = new TestChainhookPayloadBuilder()
-          .apply()
-          .block({ height: blockHeights.next().value })
-          .transaction({ hash: txHashTransfer })
-          .inscriptionRevealed(
-            brc20Reveal({
-              json: {
-                p: 'brc-20',
-                op: 'transfer',
-                tick: 'ABCD',
-                amt: '1000',
-              },
-              number,
-              ordinal_number: number,
-              tx_id: txHashTransfer,
-              address: addressB,
-            })
-          )
-          .build();
-        await db.updateInscriptions(payloadTransfer);
-        // (send inscription, transfer_send)
-        const txHashTransferSend = randomHash();
-        const payloadTransferSend = new TestChainhookPayloadBuilder()
-          .apply()
-          .block({ height: blockHeights.next().value })
-          .transaction({ hash: txHashTransferSend })
-          .inscriptionTransferred({
-            ordinal_number: number,
-            destination: { type: 'transferred', value: addressA },
-            satpoint_pre_transfer: `${txHashTransfer}:0:0`,
-            satpoint_post_transfer: `${txHashTransferSend}:0:0`,
-            post_transfer_output_value: null,
-            tx_index: 0,
-          })
-          .build();
-        await db.updateInscriptions(payloadTransferSend);
-
-        let response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens`,
-        });
-        expect(response.statusCode).toBe(200);
-        let json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toHaveLength(2);
-
-        // WITHOUT tx_count sort:
-        expect(json.results).toEqual([
-          // The first result is the token with the latest activity (ABCD)
-          expect.objectContaining({
-            ticker: 'ABCD',
-            tx_count: 4,
-          } as Brc20TokenResponse),
-          expect.objectContaining({
-            ticker: 'PEPE',
-            tx_count: 11,
-          } as Brc20TokenResponse),
-        ]);
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens?order_by=tx_count`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toHaveLength(2);
-
-        // WITH tx_count sort: The first result is the most active token (PEPE)
-        expect(json.results).toEqual([
-          expect.objectContaining({
-            ticker: 'PEPE',
-            tx_count: 11,
-          } as Brc20TokenResponse),
-          expect.objectContaining({
-            ticker: 'ABCD',
-            tx_count: 4,
-          } as Brc20TokenResponse),
-        ]);
-
-        // Rollback PEPE mints
-        for (const payload of pepeMints) {
-          const payloadRollback = { ...payload, apply: [], rollback: payload.apply };
-          await db.updateInscriptions(payloadRollback);
-        }
-
-        // WITH tx_count sort: The first result is the most active token (now ABCD)
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens?order_by=tx_count`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toHaveLength(2);
-        expect(json.results).toEqual([
-          expect.objectContaining({
-            ticker: 'ABCD',
-            tx_count: 4,
-          } as Brc20TokenResponse),
-          expect.objectContaining({
-            ticker: 'PEPE',
-            tx_count: 1, // only the deploy remains
-          } as Brc20TokenResponse),
-        ]);
-
-        // Rollback ABCD transfer
-        await db.updateInscriptions({
-          ...payloadTransferSend,
-          apply: [],
-          rollback: payloadTransferSend.apply,
-        });
-        await db.updateInscriptions({
-          ...payloadTransfer,
-          apply: [],
-          rollback: payloadTransfer.apply,
-        });
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens?order_by=tx_count`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toHaveLength(2);
-        expect(json.results).toEqual([
-          expect.objectContaining({
-            ticker: 'ABCD',
-            tx_count: 2, // only the deploy and mint remain
-          } as Brc20TokenResponse),
-          expect.objectContaining({
-            ticker: 'PEPE',
-            tx_count: 1,
-          } as Brc20TokenResponse),
-        ]);
-      });
-    });
-
-    describe('/brc-20/activity', () => {
-      test('activity for token transfers', async () => {
-        // Setup
-        const inscriptionNumbers = incrementing(0);
-        const blockHeights = incrementing(BRC20_GENESIS_BLOCK);
-        const addressA = 'bc1q6uwuet65rm6xvlz7ztw2gvdmmay5uaycu03mqz';
-        const addressB = 'bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4';
-
-        // A deploys PEPE
-        let number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE deploy is in the activity feed
-        let response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        let json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // A mints 10000 PEPE
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '10000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-            } as Brc20ActivityResponse),
-            expect.objectContaining({
-              operation: 'mint',
-              ticker: 'PEPE',
-              address: addressA,
-              mint: {
-                amount: '10000.000000000000000000',
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // B mints 10000 PEPE
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '10000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressB,
-              })
-            )
-            .build()
-        );
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(3);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'mint',
-              ticker: 'PEPE',
-              address: addressB,
-              mint: {
-                amount: '10000.000000000000000000',
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // A creates transfer of 9000 PEPE
-        const transferHash = randomHash();
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHash })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'transfer',
-                  tick: 'PEPE',
-                  amt: '9000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: transferHash,
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(4);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer',
-              ticker: 'PEPE',
-              address: addressA,
-              tx_id: transferHash,
-              transfer: {
-                amount: '9000.000000000000000000',
-                from_address: addressA,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // A sends transfer inscription to B (aka transfer/sale)
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionTransferred({
-              destination: { type: 'transferred', value: addressB },
-              tx_index: 0,
-              ordinal_number: number,
-              post_transfer_output_value: null,
-              satpoint_pre_transfer: `${transferHash}:0:0`,
-              satpoint_post_transfer:
-                '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-            })
-            .build()
-        );
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(5);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHash),
-              address: addressB,
-              transfer_send: {
-                amount: '9000.000000000000000000',
-                from_address: addressA,
-                to_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&operation=transfer_send`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHash),
-              address: addressB,
-              transfer_send: {
-                amount: '9000.000000000000000000',
-                from_address: addressA,
-                to_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-      });
-
-      test('activity for multiple token transfers among three participants', async () => {
-        // Step 1: A deploys a token
-        // Step 2: A mints 1000 of the token
-        // Step 3: B mints 2000 of the token
-        // Step 4: A creates a transfer to B
-        // Step 5: B creates a transfer to C
-        // Step 6: A transfer_send the transfer to B
-        // Step 7: B transfer_send the transfer to C
-
-        // Setup
-        const inscriptionNumbers = incrementing(0);
-        const blockHeights = incrementing(BRC20_GENESIS_BLOCK);
-        const addressA = 'bc1q6uwuet65rm6xvlz7ztw2gvdmmay5uaycu03mqz';
-        const addressB = 'bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4';
-        const addressC = 'bc1q9d80h0q5d3f54w7w8c3l2sguf9uset4ydw9xj2';
-
-        // Step 1: A deploys a token
-        let number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE deploy is in the activity feed
-        let response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        let json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 2: A mints 1000 of the token
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '1000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE mint is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'mint',
-              ticker: 'PEPE',
-              address: addressA,
-              mint: {
-                amount: '1000.000000000000000000',
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressA}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-            expect.objectContaining({
-              operation: 'mint',
-              ticker: 'PEPE',
-              address: addressA,
-              mint: {
-                amount: '1000.000000000000000000',
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 3: B mints 2000 of the token
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '2000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressB,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE mint is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(3);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'mint',
-              ticker: 'PEPE',
-              address: addressB,
-              mint: {
-                amount: '2000.000000000000000000',
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 4: A creates a transfer to B
-        const transferHashAB = randomHash();
-        const numberAB = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHashAB })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'transfer',
-                  tick: 'PEPE',
-                  amt: '1000',
-                },
-                number: numberAB,
-                ordinal_number: numberAB,
-                tx_id: transferHashAB,
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE transfer is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(4);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer',
-              ticker: 'PEPE',
-              address: addressA,
-              tx_id: transferHashAB,
-              transfer: {
-                amount: '1000.000000000000000000',
-                from_address: addressA,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressA}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(3);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer',
-              ticker: 'PEPE',
-              address: addressA,
-              tx_id: transferHashAB,
-              transfer: {
-                amount: '1000.000000000000000000',
-                from_address: addressA,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 5: B creates a transfer to C
-        const transferHashBC = randomHash();
-        const numberBC = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHashBC })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'transfer',
-                  tick: 'PEPE',
-                  amt: '2000',
-                },
-                number: numberBC,
-                ordinal_number: numberBC,
-                tx_id: transferHashBC,
-                address: addressB,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE transfer is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(5);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer',
-              ticker: 'PEPE',
-              address: addressB,
-              tx_id: transferHashBC,
-              transfer: {
-                amount: '2000.000000000000000000',
-                from_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 6: A transfer_send the transfer to B
-        const transferHashABSend = randomHash();
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHashABSend })
-            .inscriptionTransferred({
-              destination: { type: 'transferred', value: addressB },
-              tx_index: 0,
-              ordinal_number: numberAB,
-              post_transfer_output_value: null,
-              satpoint_pre_transfer: `${transferHashAB}:0:0`,
-              satpoint_post_transfer: `${transferHashABSend}:0:0`,
-            })
-            .build()
-        );
-        // A gets the transfer send in its feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressA}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(4);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashAB),
-              address: addressB,
-              transfer_send: {
-                amount: '1000.000000000000000000',
-                from_address: addressA,
-                to_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-        // B gets the transfer send in its feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressB}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(3);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashAB),
-              address: addressB,
-              transfer_send: {
-                amount: '1000.000000000000000000',
-                from_address: addressA,
-                to_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Verify that the PEPE transfer_send is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(6);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashAB),
-              address: addressB,
-              transfer_send: {
-                amount: '1000.000000000000000000',
-                from_address: addressA,
-                to_address: addressB,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 7: B transfer_send the transfer to C
-        const transferHashBCSend = randomHash();
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: transferHashBCSend })
-            .inscriptionTransferred({
-              destination: { type: 'transferred', value: addressC },
-              tx_index: 0,
-              ordinal_number: numberBC,
-              post_transfer_output_value: null,
-              satpoint_pre_transfer: `${transferHashBC}:0:0`,
-              satpoint_post_transfer: `${transferHashBCSend}:0:0`,
-            })
-            .build()
-        );
-
-        // Verify that the PEPE transfer_send is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(7);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashBC),
-              address: addressC,
-              transfer_send: {
-                amount: '2000.000000000000000000',
-                from_address: addressB,
-                to_address: addressC,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-        // B gets the transfer send in its feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressB}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(4);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashBC),
-              address: addressC,
-              transfer_send: {
-                amount: '2000.000000000000000000',
-                from_address: addressB,
-                to_address: addressC,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-        // C gets the transfer send in its feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEPE&address=${addressC}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'transfer_send',
-              ticker: 'PEPE',
-              tx_id: expect.not.stringMatching(transferHashBC),
-              address: addressC,
-              transfer_send: {
-                amount: '2000.000000000000000000',
-                from_address: addressB,
-                to_address: addressC,
-              },
-            } as Brc20ActivityResponse),
-          ])
-        );
-      });
-
-      test('activity for multiple token creation', async () => {
-        const inscriptionNumbers = incrementing(0);
-        const blockHeights = incrementing(BRC20_GENESIS_BLOCK);
-        const addressA = 'bc1q6uwuet65rm6xvlz7ztw2gvdmmay5uaycu03mqz';
-
-        // Step 1: Create a token PEPE
-        let number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '21000000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEPE deploy is in the activity feed
-        let response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity`,
-        });
-        expect(response.statusCode).toBe(200);
-        let json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Step 2: Create a token PEER
-        number = inscriptionNumbers.next().value;
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({ height: blockHeights.next().value })
-            .transaction({ hash: randomHash() })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEER',
-                  max: '21000000',
-                },
-                number,
-                ordinal_number: number,
-                tx_id: randomHash(),
-                address: addressA,
-              })
-            )
-            .build()
-        );
-
-        // Verify that the PEER deploy is in the activity feed
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEER',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-          ])
-        );
-
-        // Verify that no events are available before the first block height
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?ticker=PEER&block_height=${BRC20_GENESIS_BLOCK}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(0);
-        expect(json.results).toEqual([]);
-
-        // Verify that the PEER deploy is not in the activity feed when using block_height parameter
-        response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/activity?block_height=${BRC20_GENESIS_BLOCK}`,
-        });
-        expect(response.statusCode).toBe(200);
-        json = response.json();
-        expect(json.total).toBe(1);
-        expect(json.results).toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              operation: 'deploy',
-              ticker: 'PEPE',
-              address: addressA,
-              deploy: expect.objectContaining({
-                max_supply: '21000000.000000000000000000',
-              }),
-            } as Brc20ActivityResponse),
-          ])
-        );
-        // Should NOT include PEER at this block height
-        expect(json.results).not.toEqual(
-          expect.arrayContaining([
-            expect.objectContaining({
-              ticker: 'PEER',
-            } as Brc20ActivityResponse),
-          ])
-        );
-      });
-    });
-
-    describe('/brc-20/token/holders', () => {
-      test('displays holders for token', async () => {
-        const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
-        await deployAndMintPEPE(address);
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({
-              height: BRC20_GENESIS_BLOCK + 2,
-              hash: '0000000000000000000034dd2daec375371800da441b17651459b2220cbc1a6e',
-            })
-            .transaction({
-              hash: '633648e0e1ddcab8dea0496a561f2b08c486ae619b5634d7bb55d7f0cd32ef16',
-            })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'mint',
-                  tick: 'PEPE',
-                  amt: '2000',
-                },
-                number: 2,
-                ordinal_number: 2,
-                tx_id: '633648e0e1ddcab8dea0496a561f2b08c486ae619b5634d7bb55d7f0cd32ef16',
-                address: 'bc1qp9jgp9qtlhgvwjnxclj6kav6nr2fq09c206pyl',
-              })
-            )
-            .build()
-        );
-
-        const response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens/PEPE/holders`,
-        });
-        expect(response.statusCode).toBe(200);
-        const json = response.json();
-        expect(json.total).toBe(2);
-        expect(json.results).toStrictEqual([
-          {
-            address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-            overall_balance: '10000.000000000000000000',
-          },
-          {
-            address: 'bc1qp9jgp9qtlhgvwjnxclj6kav6nr2fq09c206pyl',
-            overall_balance: '2000.000000000000000000',
-          },
-        ]);
-      });
-
-      test('shows empty list on token with no holders', async () => {
-        await db.updateInscriptions(
-          new TestChainhookPayloadBuilder()
-            .apply()
-            .block({
-              height: BRC20_GENESIS_BLOCK,
-              hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-            })
-            .transaction({
-              hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-            })
-            .inscriptionRevealed(
-              brc20Reveal({
-                json: {
-                  p: 'brc-20',
-                  op: 'deploy',
-                  tick: 'PEPE',
-                  max: '250000',
-                },
-                number: 0,
-                ordinal_number: 0,
-                tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-                address: 'bc1qp9jgp9qtlhgvwjnxclj6kav6nr2fq09c206pyl',
-              })
-            )
-            .build()
-        );
-        const response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens/PEPE/holders`,
-        });
-        expect(response.statusCode).toBe(200);
-        const json = response.json();
-        expect(json.total).toBe(0);
-        expect(json.results).toStrictEqual([]);
-      });
-
-      test('shows 404 on token not found', async () => {
-        const response = await fastify.inject({
-          method: 'GET',
-          url: `/ordinals/brc-20/tokens/PEPE/holders`,
-        });
-        expect(response.statusCode).toBe(404);
-      });
-    });
-  });
-
   describe('rollbacks', () => {
     test('reflects rollbacks on balances and counts correctly', async () => {
       const address = 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td';
       const address2 = '3QNjwPDRafjBm9XxJpshgk3ksMJh3TFxTU';
-      await deployAndMintPEPE(address);
+      await deployAndMintPEPE(db, address);
 
       // Transfer and send PEPE
       const transferPEPE = new TestChainhookPayloadBuilder()
@@ -4172,20 +1012,14 @@ describe('BRC-20', () => {
         .transaction({
           hash: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
         })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'transfer',
-              tick: 'PEPE',
-              amt: '9000',
-            },
-            number: 2,
-            ordinal_number: 2,
-            tx_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a',
-            address: address,
-          })
-        )
+        .brc20({
+          transfer: {
+            inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+            tick: 'PEPE',
+            address,
+            amt: '9000',
+          },
+        })
         .build();
       await db.updateInscriptions(transferPEPE);
       const sendPEPE = new TestChainhookPayloadBuilder()
@@ -4197,15 +1031,14 @@ describe('BRC-20', () => {
         .transaction({
           hash: '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac',
         })
-        .inscriptionTransferred({
-          ordinal_number: 2,
-          destination: { type: 'transferred', value: address2 },
-          satpoint_pre_transfer:
-            'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47a:0:0',
-          satpoint_post_transfer:
-            '7edaa48337a94da327b6262830505f116775a32db5ad4ad46e87ecea33f21bac:0:0',
-          post_transfer_output_value: null,
-          tx_index: 0,
+        .brc20({
+          transfer_send: {
+            inscription_id: 'eee52b22397ea4a4aefe6a39931315e93a157091f5a994216c0aa9c8c6fef47ai0',
+            tick: 'PEPE',
+            amt: '9000',
+            sender_address: address,
+            receiver_address: address2,
+          },
         })
         .build();
       await db.updateInscriptions(sendPEPE);
@@ -4219,20 +1052,17 @@ describe('BRC-20', () => {
         .transaction({
           hash: '8354e85e87fa2df8b3a06ec0b9d395559b95174530cb19447fc4df5f6d4ca84d',
         })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'deploy',
-              tick: '🔥',
-              max: '1000',
-            },
-            number: 3,
-            ordinal_number: 3,
-            tx_id: '8354e85e87fa2df8b3a06ec0b9d395559b95174530cb19447fc4df5f6d4ca84d',
-            address: address,
-          })
-        )
+        .brc20({
+          deploy: {
+            inscription_id: '8354e85e87fa2df8b3a06ec0b9d395559b95174530cb19447fc4df5f6d4ca84di0',
+            tick: '🔥',
+            max: '1000',
+            lim: '1000',
+            dec: '18',
+            address,
+            self_mint: false,
+          },
+        })
         .build();
       await db.updateInscriptions(deployFIRE);
       const mintFIRE = new TestChainhookPayloadBuilder()
@@ -4244,20 +1074,14 @@ describe('BRC-20', () => {
         .transaction({
           hash: '81f4ee2c247c5f5c0d3a6753fef706df410ea61c2aa6d370003b98beb041b887',
         })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'mint',
-              tick: '🔥',
-              amt: '500',
-            },
-            number: 4,
-            ordinal_number: 4,
-            tx_id: '81f4ee2c247c5f5c0d3a6753fef706df410ea61c2aa6d370003b98beb041b887',
-            address: address,
-          })
-        )
+        .brc20({
+          mint: {
+            inscription_id: '81f4ee2c247c5f5c0d3a6753fef706df410ea61c2aa6d370003b98beb041b887i0',
+            tick: '🔥',
+            address,
+            amt: '500',
+          },
+        })
         .build();
       await db.updateInscriptions(mintFIRE);
       // Transfer and send 🔥 to self
@@ -4270,20 +1094,14 @@ describe('BRC-20', () => {
         .transaction({
           hash: 'c1c7f1d5c10a30605a8a5285ca3465a4f75758ed9b7f201e5ef62727e179966f',
         })
-        .inscriptionRevealed(
-          brc20Reveal({
-            json: {
-              p: 'brc-20',
-              op: 'transfer',
-              tick: '🔥',
-              amt: '100',
-            },
-            number: 5,
-            ordinal_number: 5,
-            tx_id: 'c1c7f1d5c10a30605a8a5285ca3465a4f75758ed9b7f201e5ef62727e179966f',
-            address: address,
-          })
-        )
+        .brc20({
+          transfer: {
+            inscription_id: 'c1c7f1d5c10a30605a8a5285ca3465a4f75758ed9b7f201e5ef62727e179966fi0',
+            tick: '🔥',
+            address,
+            amt: '100',
+          },
+        })
         .build();
       await db.updateInscriptions(transferFIRE);
       const sendFIRE = new TestChainhookPayloadBuilder()
@@ -4295,15 +1113,14 @@ describe('BRC-20', () => {
         .transaction({
           hash: 'a00d01a3e772ce2219ddf3fe2fe4053be071262d9594f11f018fdada7179ae2d',
         })
-        .inscriptionTransferred({
-          ordinal_number: 5,
-          destination: { type: 'transferred', value: address }, // To self
-          satpoint_pre_transfer:
-            'c1c7f1d5c10a30605a8a5285ca3465a4f75758ed9b7f201e5ef62727e179966f:0:0',
-          satpoint_post_transfer:
-            'a00d01a3e772ce2219ddf3fe2fe4053be071262d9594f11f018fdada7179ae2d:0:0',
-          post_transfer_output_value: null,
-          tx_index: 0,
+        .brc20({
+          transfer_send: {
+            tick: '🔥',
+            inscription_id: 'c1c7f1d5c10a30605a8a5285ca3465a4f75758ed9b7f201e5ef62727e179966fi0',
+            amt: '100',
+            sender_address: address,
+            receiver_address: address,
+          },
         })
         .build();
       await db.updateInscriptions(sendFIRE);
